# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vocab Trainer API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/review/queue": {
      "get": {
        "tags": ["review"],
        "summary": "Review Queue",
        "operationId": "review_queue_review_queue_get",
        "parameters": [
          {
            "name": "deck_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardOut"
                  },
                  "title": "Response Review Queue Review Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/review": {
      "post": {
        "tags": ["review"],
        "summary": "Record Review",
        "operationId": "record_review_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/review/summary": {
      "get": {
        "tags": ["review"],
        "summary": "Review Summary",
        "operationId": "review_summary_review_summary_get",
        "parameters": [
          {
            "name": "deck_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/review/skills/queue": {
      "get": {
        "tags": ["review"],
        "summary": "Skill Review Queue",
        "operationId": "skill_review_queue_review_skills_queue_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurriculumSkillReviewOut"
                  },
                  "title": "Response Skill Review Queue Review Skills Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/review/skills": {
      "post": {
        "tags": ["review"],
        "summary": "Record Skill Review",
        "operationId": "record_skill_review_review_skills_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumSkillItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CardOut": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "word": {
            "type": "string",
            "title": "Word"
          },
          "translation": {
            "type": "string",
            "title": "Translation"
          },
          "transcription": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcription"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "due": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due"
          }
        },
        "type": "object",
        "required": ["id", "word", "translation", "transcription"],
        "title": "CardOut"
      },
...
      "CurriculumSkillItemOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "module_id": {
            "type": "string",
            "title": "Module Id"
          },
          "source_item_id": {
            "type": "string",
            "title": "Source Item Id"
          },
          "is_leech": {
            "type": "boolean",
            "title": "Is Leech"
          }
        },
        "type": "object",
        "required": ["id", "skill", "module_id", "source_item_id", "is_leech"],
        "title": "CurriculumSkillItemOut"
      },
      "CurriculumSkillReviewOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "module_id": {
            "type": "string",
            "title": "Module Id"
          },
          "source_item_id": {
            "type": "string",
            "title": "Source Item Id"
          },
          "is_leech": {
            "type": "boolean",
            "title": "Is Leech"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "answers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Answers",
            "default": []
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "id",
          "skill",
          "module_id",
          "source_item_id",
          "is_leech",
          "type",
          "prompt",
          "explanation"
        ],
        "title": "CurriculumSkillReviewOut",
        "description": "A due skill item with its source quiz item for the review UI."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewIn": {
        "properties": {
          "card_id": {
            "type": "integer",
            "title": "Card Id"
          },
          "rating": {
            "type": "integer",
            "enum": [1, 2, 3, 4],
            "title": "Rating"
          }
        },
        "type": "object",
        "required": ["card_id", "rating"],
        "title": "ReviewIn"
      },
      "ReviewSummaryOut": {
        "properties": {
          "next_due": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Due"
          },
          "reviewed_today": {
            "type": "integer",
            "title": "Reviewed Today"
          }
        },
        "type": "object",
        "required": ["next_due", "reviewed_today"],
        "title": "ReviewSummaryOut"
      },
...
      "SkillReviewIn": {
        "properties": {
          "skill_item_id": {
            "type": "integer",
            "title": "Skill Item Id"
          },
          "rating": {
            "type": "integer",
            "enum": [1, 2, 3, 4],
            "title": "Rating"
          }
        },
        "type": "object",
        "required": ["skill_item_id", "rating"],
        "title": "SkillReviewIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}